Accept missing HTTP status code during log parsing
As it turns out, the HTTP status code field may also be blank in the
access logs from S3. Update the parser to reflect that, and add a
real-world test case that was failing.

+++ aggregator/process_raw.py
@@ -24,7 +24,7 @@ column_types = {
     "operation": "object",
     "key": "object",
     "request_uri": "object",
-    "http_status": "int64",
+    "http_status": "float64",
     "error_code": "object",
     "bytes_sent": "float64",
     "object_size": "float64",
